requirements.txt
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
Pillow>=10.2.0
gspread>=6.0.0
google-auth>=2.27.0
